Feat[Contributors]: Refactor code
Delete Unecessary code, and correct spellings, make hospitals card work
as it was in the 2nd commit of master, refactor code.
