+++ src/components/Hospitals/Hospitals.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/styles';
 import HospitalCard from '../HospitalCard/HospitalCard';
+import {hospitals} from 'data.js';
 import {Grid} from '@material-ui/core';
 const useStyles = makeStyles(theme => ({
   content: {
@@ -18,7 +19,6 @@ const useStyles = makeStyles(theme => ({
 
 
 const Hospitals = props => {
-  const { hospitals } = props;
   const classes = useStyles();
 
   return (
