+++ src/components/index.js
@@ -1,3 +1,2 @@
 export { default as HospitalCard } from './HospitalCard';
 export { default as Hospitals } from './Hospitals';
-export { default as Contact } from './Contact';
