--- src/views/Contact/index.js
@@ -1,1 +1,0 @@
-export { default } from './Contact';
