+++ src/views/index.js
@@ -1,2 +1,2 @@
 export {default as LandingPage} from './LandingPage';
-export {default as Contact} from './Contact';
+export {default as ContactPage} from './ContactPage';
